+++ eql/__init__.py
@@ -66,7 +66,7 @@ from .walkers import (
     Walker,
 )
 
-__version__ = '0.9.15'
+__version__ = '0.9.16'
 __all__ = (
     "__version__",
     "AnalyticOutput",